Moved Firebase functions to /pages/services
